+++ packages/core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @shallot-ui/core
 
+## 5.0.14
+
+### Patch Changes
+
+- Adds modal component
+- Updated dependencies
+  - @shallot-ui/theme@0.27.14
+
 ## 5.0.13
 
 ### Patch Changes